Add support in application UI for resetting model
This change makes it possible to reset an autoencoder model on the fly from the UI. Previously, this behaviour required a restart of the application.

Note that a reset causes a random initialization of weights, i.e. the reset does not revert to the initial state.

This change also generalizes AutoEncoder.add_noise to support more than single-layered networks.