Fix recall exception bug
Before this change, a list index exception could occur during recal in applications. The problem was that int(random.uniform(...)) was used in a risky way, expecting that the max end-point would not be included. But according to the python documentation, the max end-point "may or may not be included in the range depending on floating-point rounding".

After the change, random.randint is used instead.