Don't freeze application when saving recording
Before this change, when a recording was stopped and therefore saved, the whole application frooze until the saving was complete. This could take a very long time for long recordings.

After this change, the saving takes place in a separate thread, allowing the application to continue to run. Furthermore, a new recording can start while the previous is still saving.