Fix recording bug in applications
After the optimization in c4263d5c21a623f3a6bd5bdbd686f5d21b91175d, the saved recording didn't contain changes in constraints. The reason was that the frames were exported with a default entity, rather than the actual entity with potentially modified contraints.

This change solves the problem. Some of the optimization is lost.